feat: capture options from contract decorator as class property
