+++ packages/algo-ts/src/impl/primitives.ts
@@ -1,4 +1,4 @@
-import type { biguint, BigUintCompat, bytes, BytesBacked, BytesCompat, uint64, Uint64Compat } from '../index'
+import type { biguint, BigUintCompat, bytes, BytesCompat, uint64, Uint64Compat } from '../index'
 import { base32ToUint8Array } from './base-32'
 import {
   base64ToUint8Array,
@@ -29,21 +29,6 @@ export function toExternalValue(val: uint64 | biguint | bytes | string) {
   if (instance instanceof Uint64Cls) return instance.asBigInt()
   if (instance instanceof BigUintCls) return instance.asBigInt()
   if (typeof val === 'string') return val
-}
-export const toBytes = (val: unknown): bytes => {
-  if (val instanceof AlgoTsPrimitiveCls) return val.toBytes().asAlgoTs()
-  if ((val as BytesBacked).bytes) return BytesCls.fromCompat((val as BytesBacked).bytes).asAlgoTs()
-
-  switch (typeof val) {
-    case 'string':
-      return BytesCls.fromCompat(val).asAlgoTs()
-    case 'bigint':
-      return BigUintCls.fromCompat(val).toBytes().asAlgoTs()
-    case 'number':
-      return Uint64Cls.fromCompat(val).toBytes().asAlgoTs()
-    default:
-      internalError(`Unsupported arg type ${nameOfType(val)}`)
-  }
 }
 
 /**
