+++ packages/algo-ts/src/internal.ts
@@ -1,3 +1,4 @@
+export { ContractOptionsSymbol } from './base-contract'
 export { ExecutionContext, ctxMgr } from './execution-context'
 export * as encodingUtil from './impl/encoding-util'
 export * as errors from './impl/errors'
